requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Migrate
email-validator
Flask-Mail
gspread
oauth2client
python-dotenv
Werkzeug
WTForms
pandas
openpyxl
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
numpy
opencv-python-headless
pytz
insightface
onnxruntime